+++ .vscode/cspell.json
@@ -75,6 +75,13 @@
         "msal",
         "imds"
       ]
+    },
+    {
+      "filename": "sdk/typespec/**",
+      "flagWords": [
+        "azure",
+        "azurite"
+      ]
     }
   ]
 }
