+++ sdk/core/azure_core/Cargo.toml
@@ -14,6 +14,8 @@ edition.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+typespec = { workspace = true, features = ["http", "json"] }
+typespec_client_core = { workspace = true, features = ["http", "json"] }
 async-trait = { workspace = true }
 base64 = { workspace = true }
 bytes = { workspace = true }
