+++ sdk/core/azure_core/src/error.rs
@@ -1,6 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#![doc = include_str!("../README.md")]
-
-pub mod error;
+pub use typespec::error::*;
+pub use typespec_client_core::error::HttpError;
