+++ sdk/identity/azure_identity/src/refresh_token.rs
@@ -10,6 +10,7 @@ use azure_core::{json::from_json, Method};
 use serde::Deserialize;
 use std::fmt;
 use std::sync::Arc;
+use typespec_client_core::error::http_response_from_body;
 use url::form_urlencoded;
 
 /// Exchange a refresh token for a new access token and refresh token
@@ -51,8 +52,8 @@ pub async fn exchange(
         rsp.json().await.map_kind(ErrorKind::Credential)
     } else {
         let rsp_body = rsp.into_body().collect().await?;
-        let token_error: RefreshTokenError = from_json(&rsp_body)
-            .map_err(|_| ErrorKind::http_response_from_body(rsp_status, &rsp_body))?;
+        let token_error: RefreshTokenError =
+            from_json(&rsp_body).map_err(|_| http_response_from_body(rsp_status, &rsp_body))?;
         Err(Error::new(ErrorKind::Credential, token_error))
     }
 }
