+++ sdk/typespec/Cargo.toml
@@ -3,14 +3,23 @@ name = "typespec"
 version = "0.1.0"
 authors.workspace = true
 edition.workspace = true
-description = "Project root for all TypeSpec-related crates"
+description = "Project root for all TypeSpec-related crates."
 homepage = "https://typespec.io"
 repository.workspace = true
 license.workspace = true
-categories = [
-  "compilers",
-  "development-tools",
-]
-keywords = [
-  "typespec",
-]
+categories = ["compilers", "development-tools"]
+keywords = ["typespec"]
+
+[dependencies]
+base64 = { workspace = true }
+http-types = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true }
+url = { workspace = true }
+
+[dev-dependencies]
+thiserror = { workspace = true }
+
+[features]
+default = ["http", "json"]
+http = ["dep:http-types"]
+json = ["dep:serde_json"]
