+++ sdk/typespec/typespec_client_core/README.md
@@ -1,0 +1,3 @@
+# TypeSpec Client Runtime
+
+This is the runtime for [TypeSpec](https://typespec.io)-generated clients.
