+++ sdk/typespec/typespec_client_core/src/http/mod.rs
@@ -1,6 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#![doc = include_str!("../README.md")]
+pub mod headers;
+mod response;
 
-pub mod error;
+pub use response::*;
+
+pub use http_types::StatusCode;
