+++ sdk/typespec/typespec_client_core/src/http/response.rs
@@ -1,18 +1,20 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 use crate::{
-    error::{ErrorKind, ResultExt},
-    headers::Headers,
+    http::{headers::Headers, StatusCode},
     json::from_json,
-    StatusCode,
 };
 use bytes::Bytes;
 use futures::{Stream, StreamExt};
 use serde::de::DeserializeOwned;
 use std::{fmt, marker::PhantomData, pin::Pin};
+use typespec::error::{ErrorKind, ResultExt};
 
 #[cfg(not(target_arch = "wasm32"))]
-pub(crate) type PinnedStream = Pin<Box<dyn Stream<Item = crate::Result<Bytes>> + Send + Sync>>;
+pub type PinnedStream = Pin<Box<dyn Stream<Item = crate::Result<Bytes>> + Send + Sync>>;
 #[cfg(target_arch = "wasm32")]
-pub(crate) type PinnedStream = Pin<Box<dyn Stream<Item = crate::Result<Bytes>>>>;
+pub type PinnedStream = Pin<Box<dyn Stream<Item = crate::Result<Bytes>>>>;
 
 /// An HTTP response.
 pub struct RawResponse {
