+++ sdk/typespec/typespec_client_core/src/lib.rs
@@ -4,3 +4,9 @@
 #![doc = include_str!("../README.md")]
 
 pub mod error;
+#[cfg(feature = "http")]
+pub mod http;
+#[cfg(feature = "json")]
+pub mod json;
+
+pub use typespec::error::*;
